pyproject.toml
[project]
name = "dci-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "gql>=4.0.0",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.33",
    "openai>=1.107.2",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "requests>=2.32.5",
    "requests-toolbelt>=1.0.0",
    "tenacity>=9.1.2",
]
